requirements.txt
asgiref==3.4.1
astroid==2.8.2
autopep8==1.5.7
Django==3.2.8
django-environ==0.7.0
djangorestframework==3.12.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.0.3
platformdirs==2.4.0
pycodestyle==2.8.0
pylint==2.11.1
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1
